extract components for <TinaMarkdown> into one dir
