+++ components/blocks/TextAndMediaColumns.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { VideoGridComponent } from './MediaComponent'
 import { TinaMarkdown } from 'tinacms/dist/rich-text'
-import { Components } from 'tinacms/dist/rich-text'
+import { textAndMediaColumnsComponentTemplate } from './TextAndMediaColumns.template'
 
 const TextAndMediaColumnsComponent = ({ data }) => {
     const { mediaColumnItem, body, isVideoOnLeft } = data || {}
@@ -11,19 +11,6 @@ const TextAndMediaColumnsComponent = ({ data }) => {
         youtube: 'PageBlocksTextMediaColumnComponentMediaColumnItemYoutubeMedia'
     }
 
-
-
-    const components: Components<{}> = {
-        p: (props) => <>
-            {props.children.props.content.map(content => <p className="text-lg lg:text-xl text-black"> {content.text} </p>)}
-            <br/>
-        </>,
-        h6: (props) => <>
-            {props.children.props.content.map(content => <h6 className="font-tuner text-3xl lg:text-4xl lg:leading-tight text-orange-400"> {content.text} </h6>)}
-            <br />
-        </>
-    }
-
     return (
         <div className="px-3 max-w-screen-xl mx-auto pb-4 pt-16 w-full">
             <div className="lg:grid lg:grid-cols-7 mx-auto md:pb-16">
@@ -31,7 +18,7 @@ const TextAndMediaColumnsComponent = ({ data }) => {
                     {VideoGridComponent({ data: { mediaItem: mediaColumnItem, typenames } })}
                 </div>
                 <div className={`lg:col-span-3 py-12 lg:py-0 mx-12 md:mx-16 text-center lg:text-start lg:mx-30 my-auto relative row-start-1 ${isVideoOnLeft ? "lg:col-start-5" : "lg:col-start-1"}`}>
-                    {body && <TinaMarkdown content={body} components={components}/>}
+                    {body && <TinaMarkdown content={body} components={textAndMediaColumnsComponentTemplate}/>}
                     
                 </div>
             </div>
