+++ components/tinaMarkdownComponents/contentComponents.tsx
@@ -1,0 +1,9 @@
+import { Actions } from "components/blocks";
+import { NewsletterBlock } from "components/blocks/Newsletter";
+import { SocialBlock } from "components/blocks/Social";
+
+export const contentComponents = {
+    actions: Actions,
+    social: SocialBlock,
+    newsletter: NewsletterBlock,
+  }
