+++ components/tinaMarkdownComponents/pricingComponents.tsx
@@ -1,0 +1,4 @@
+export const pricingComponents = {
+    p: (props) => <p className="text-xl" {...props} />,
+    strong: (props) => <strong className="font-bold text-xl" {...props} />,
+  }
